Use int durations; add Markov generator and tests
Change duration/tiempos types from float to int across DTO/model classes to standardize durations as integer milliseconds (EstadisticasDeTransicion, EventoMusical, TransicionMarkov). Add services: GeneradorMatrizMarkov (builds a MatrizMarkov from grouped transition data) and GeneradorProcedural (generates SecuenciaMusical procedurally using Markov transitions and random selection of durations). Add tests (TestGeneradorProcedural.py, TestMatrixGeneration.py) with mocked transition data to exercise matrix generation and sequence generation. The new generator implementations rely on MatrizMarkov, NotaMusical, TransicionMarkov and DataTransicionAgrupada and implement probability calculation and random interval-based transition selection.

+++ dto/EstadisticasDeTransicion.py
@@ -3,4 +3,4 @@ from dataclasses import dataclass
 @dataclass
 class EstadisticasDeTransicion:
     _conteo:int
-    _tiempos:list[float]
+    _tiempos:list[int]
